+++ build_runner/lib/src/generate/watch_impl.dart
@@ -92,6 +92,13 @@ Future<ServeHandler> watch(
   final buildPhases = await createBuildPhases(
       targetGraph, builders, builderConfigOverrides, isReleaseBuild ?? false);
 
+  if (buildPhases.isEmpty) {
+    _logger.severe('Nothing can be built, yet a build was requested.');
+    await terminator.cancel();
+    await options.logListener.cancel();
+    return null;
+  }
+
   var watch =
       _runWatch(environment, options, buildPhases, terminator.shouldTerminate);