+++ build_runner/CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 0.8.7
 
+- Now fail when a build is requested which does not build anything.
 - Clean up some error messages.
 
 ## 0.8.6